label dtype

+++ pets/utils/data.py
@@ -197,8 +197,8 @@ def parse_text_file(filename: str) -> pl.DataFrame:
     columns = ["image_id", "class_id", "species_id", "breed_id"]
     df = pl.read_csv(filename, separator=" ", has_header=False, new_columns=columns)
     df = df.with_columns(
-        (pl.col("class_id") - 1).alias("class_id").cast(pl.Int32),
-        (pl.col("species_id") - 1).alias("species_id").cast(pl.Int32),
+        (pl.col("class_id") - 1).alias("class_id"),
+        (pl.col("species_id") - 1).alias("species_id"),
     )
     return df
